alert working on delete button, Hard Mode

+++ server/public/scripts/client.js
@@ -9,11 +9,12 @@ todoList.controller('TodoController', ['$http', function($http){
 
     todo.todoArray = [];
 
+    //POST
     todo.addTask = function(newTask) {
         console.log('inside addTask', newTask);
         $http({
             method: 'POST',
-            url: '/todo',
+            url: '/todo',  
             data: newTask
         }).then(function(response){
             console.log('POST response', response);
@@ -23,6 +24,7 @@ todoList.controller('TodoController', ['$http', function($http){
         });
     }
     
+    //GET
     todo.getTodo = function () {
         $http({  
             method: 'GET',
@@ -36,6 +38,7 @@ todoList.controller('TodoController', ['$http', function($http){
     }
     todo.getTodo();
 
+    //PUT
     todo.taskDone = function (taskId, status) {
         console.log('client.js PUT mark task completed function is started', taskId);
         $http({
@@ -50,18 +53,47 @@ todoList.controller('TodoController', ['$http', function($http){
         });
     }
 
-    todo.taskDelete = function (taskId) {
-        console.log('client.js DELETE task function is started', taskId);
-        $http({
-            method: 'DELETE',
-            url: '/todo/' + taskId,
-            data: { taskDone: status }
-        }).then(function (response) {
-            console.log('Task deleted', response);
-            todo.getTodo();
-        }).catch(function (error) {
-            console.log('Error', error);
-        });
+    //DELETE vanilla-style
+    // todo.taskDelete = function (taskId) {
+    //     console.log('client.js DELETE task function is started', taskId);
+    //     $http({
+    //         method: 'DELETE',
+    //         url: '/todo/' + taskId,
+    //         data: { taskDone: status }
+    //     }).then(function (response) {
+    //         console.log('Task deleted', response);
+    //         todo.getTodo();
+    //     }).catch(function (error) {
+    //         console.log('Error', error);
+    //     });
+    // }
+
+    //DELETE
+    todo.taskDelete  = function (taskId) {
+        swal({
+            title: "Are you sure you want to delete?",
+            buttons: true,
+            dangerMode: true,
+        })
+            .then((value) => {
+                if (value === true) {
+                    swal("Task has been deleted!", {
+                        icon: "success"
+                    })
+                    $http({
+                        method: 'DELETE',
+                        url: '/todo/' + taskId
+                    }).then(function (response) {
+                        console.log('Successfully deleted: ', response);
+                        
+                    todo.getTodo();
+
+                    }).catch(function (error) {
+                        console.log('Error in deleting entry', error);
+                    })
+                }
+            });
+
     }
 
 }]);
